ImproveDetectionByConnectingParticlesOverExtendedRange
Newly implemented: a function that allows an increased range for
connecting neighbored voxels to a cilium or neighbored particles to an
intensity region. Now also diagonally adjacent voxels are linked in the
Flood Filling algorithm, if selected in the initial menu.
Bugfixes: In v0.0.5 the filtering of particles in intensity regions did
not work for all frames after frame one in the timelapse mode; now this
issue is fixed.