ImplementColocalizedOnCenterline
Introduces two new outputs per intensity channel (A, B):
Colocalized on centerline compared to BG volume [micron]: the number of
segments of the intensity profile that are higher than the automatically
determined background intensity threshold, multiplied with the stepsize
of the profile, i.e. the pixel length in the xy dimension.
Colocalized on centerline compared to BG volume [% total length]: the
number of segments of the intensity profile that are higher than the
automatically determined background intensity threshold, divided by the
total number of segments of the intensity profile and multiplied with
100.